+++ packages/terra-menu/CHANGELOG.md
@@ -9,6 +9,8 @@ Unreleased
 ### Changed
 * Minor dependency version bump
 * Removed usage of componentWillReceiveProps
+* Updated theme variables
+* Make menu item outline on focus themeable and provide browser default fallback.
 * Updated site examples to be more accessible.
 
 3.28.1 - (November 20, 2018)