+++ packages/terra-menu/CHANGELOG.md
@@ -9,6 +9,8 @@ Unreleased
 ### Changed
 * Minor dependency version bump
 * Removed usage of componentWillReceiveProps
+* Updated theme variables
+* Make menu item outline on focus themeable and provide browser default fallback.
 * Updated site examples to be more accessible.
 * Update to use new List API, create single select implementation within MenuItemGroup.